requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.8.0
pillow>=9.0.0
tqdm>=4.65.0

# YOLO
ultralytics>=8.0.0

# RF-DETR
rfdetr

# Ensemble
ensemble-boxes>=1.0.9

# Visualization and utilities
supervision>=0.16.0
matplotlib>=3.7.0
pandas>=2.0.0

# Optional: for notebook support
jupyter
ipywidgets
